+++ MacClient/session/editor/NotebookItemData.swift
@@ -6,13 +6,16 @@
 
 import Foundation
 import SyntaxParsing
+import ReactiveSwift
 
 public class NotebookItemData: NSObject {
 	@objc public let source: NSTextStorage
 	@objc public let result: NSTextStorage
 	@objc public var height: CGFloat = 40
 	public var chunk: RmdChunk
-	
+	/// this is bookkeeping for NotebookViewItem and is meant to be used by it
+	public let resultsVisible = MutableProperty(true)
+
 	public init(chunk: RmdChunk, result: String) {
 		self.chunk = chunk
 		self.source = NSTextStorage(attributedString: chunk.attributedContents)