+++ MacClient/session/editor/NotebookViewItem.swift
@@ -8,13 +8,13 @@ import Cocoa
 
 protocol NotebookViewItemDelegate: class {
 	func addChunk(after: NotebookViewItem, sender: NSButton?)
+	/// should call .resultsVisible = hide on all chunks
+	func twiddleAllChunks(hide: Bool)
 }
 
 protocol NotebookViewItem: class {
 	var delegate: NotebookViewItemDelegate? { get set }
 	var data: NotebookItemData? { get set }
 	var context: EditorContext? { get set }
-
-	func addChunk(_ sender: Any?)
 }